+++ src/main/java/com/martinsil/pvzcraft/PvZCraftClient.java
@@ -3,6 +3,8 @@ package com.martinsil.pvzcraft;
 import com.martinsil.pvzcraft.entity.ModEntities;
 import com.martinsil.pvzcraft.entity.client.PeashooterModel;
 import com.martinsil.pvzcraft.entity.client.PeashooterRenderer;
+import com.martinsil.pvzcraft.entity.client.RegularZombieModel;
+import com.martinsil.pvzcraft.entity.client.RegularZombieRenderer;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityModelLayerRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
@@ -14,6 +16,11 @@ public class PvZCraftClient implements ClientModInitializer {
     public void onInitializeClient() {
         EntityModelLayerRegistry.registerModelLayer(PeashooterModel.PEASHOOTER, PeashooterModel::getTexturedModelData);
         EntityRendererRegistry.register(ModEntities.PEASHOOTER, PeashooterRenderer::new);
+
         EntityRendererRegistry.register(ModEntities.PEA, FlyingItemEntityRenderer::new);
+
+        EntityModelLayerRegistry.registerModelLayer(RegularZombieModel.REGULAR_ZOMBIE, RegularZombieModel::getTexturedModelData);
+        EntityRendererRegistry.register(ModEntities.REGULAR_ZOMBIE, RegularZombieRenderer::new);
+
     }
 }