+++ src/main/java/com/martinsil/pvzcraft/entity/client/RegularZombieAnimations.java
@@ -1,0 +1,4 @@
+package com.martinsil.pvzcraft.entity.client;
+
+public class RegularZombieAnimations {
+}